half dtype

+++ test/testextra/test_asm_gemm.py
@@ -65,9 +65,9 @@ class TestGemmLarge(unittest.TestCase):
   def test_gemm8(self): verify_asm_gemm(1, 4096, 14336, 8192)
   def test_gemm9(self): verify_asm_gemm(8, 4096, 14336, 8192, dtype=dtypes.bfloat16, gpus=8)
   def test_gemm10(self): verify_asm_gemm(1, 4096, 8192, 4096)
-  def test_k_sharded_1(self): verify_asm_gemm_k_sharded(14336, 4096, 8*8192, dtype=dtypes.bfloat16, gpus=8)
-  def test_k_sharded_2(self): verify_asm_gemm_k_sharded(4096, 14336, 8*8192, dtype=dtypes.bfloat16, gpus=8)
-  def test_k_sharded_3(self): verify_asm_gemm_k_sharded(4096, 4096, 8*8192, dtype=dtypes.bfloat16, gpus=8)
+  def test_k_sharded_1(self): verify_asm_gemm_k_sharded(14336, 4096, 8*8192, gpus=8)
+  def test_k_sharded_2(self): verify_asm_gemm_k_sharded(4096, 14336, 8*8192, gpus=8)
+  def test_k_sharded_3(self): verify_asm_gemm_k_sharded(4096, 4096, 8*8192, gpus=8)
   def test_gemm_unsupported(self):
     with self.assertRaisesRegex(AssertionError, "shape not supported"):
       verify_asm_gemm(8, 1024, 1024, 4096, gpus=8)
